Move print_step into contextmanager
